passing java-eventsourced-counter test
