+++ samples/java-eventsourced-counter/src/test/java/com/example/actions/CounterTopicSubscriptionActionTest.java
@@ -1,33 +1,40 @@
-/* This code was generated by Akka Serverless tooling.
- * As long as this file exists it will not be re-generated.
- * You are free to make changes to this file.
- */
 package com.example.actions;
 
 import com.akkaserverless.javasdk.testkit.ActionResult;
-import com.example.actions.CounterJournalToTopicAction;
-import com.example.actions.CounterJournalToTopicActionTestKit;
 import com.example.actions.CounterTopicApi;
-import com.example.domain.CounterDomain;
-import com.google.protobuf.Any;
+import com.example.actions.CounterTopicSubscriptionAction;
+import com.example.actions.CounterTopicSubscriptionActionTestKit;
 import com.google.protobuf.Empty;
 import org.junit.Test;
 import static org.junit.Assert.*;
 
+// This class was initially generated based on the .proto definition by Akka Serverless tooling.
+//
+// As long as this file exists it will not be overwritten: you can maintain it yourself,
+// or delete it so it is regenerated as needed.
 
-public class CounterJournalToTopicActionTest {
+public class CounterTopicSubscriptionActionTest {
+
+  @Test
+  public void exampleTest() {
+    CounterTopicSubscriptionActionTestKit testKit = CounterTopicSubscriptionActionTestKit.of(CounterTopicSubscriptionAction::new);
+    // use the testkit to execute a command
+    // ActionResult<SomeResponse> result = testKit.someOperation(SomeRequest);
+    // verify the response
+    // SomeResponse actualResponse = result.getReply();
+    // assertEquals(expectedResponse, actualResponse);
+  }
 
   @Test
   public void increaseTest() {
-    CounterJournalToTopicActionTestKit testKit = CounterJournalToTopicActionTestKit.of(CounterJournalToTopicAction::new);
-    ActionResult<CounterTopicApi.Increased> result = testKit.increase(CounterDomain.ValueIncreased.newBuilder().setValue(1).build());
-    assertEquals(1, result.getReply().getValue());
+    CounterTopicSubscriptionActionTestKit testKit = CounterTopicSubscriptionActionTestKit.of(CounterTopicSubscriptionAction::new);
+    // ActionResult<Empty> result = testKit.increase(CounterTopicApi.Increased.newBuilder()...build());
   }
 
   @Test
   public void decreaseTest() {
-    CounterJournalToTopicActionTestKit testKit = CounterJournalToTopicActionTestKit.of(CounterJournalToTopicAction::new);
-    ActionResult<CounterTopicApi.Decreased> result = testKit.decrease(CounterDomain.ValueDecreased.newBuilder().setValue(1).build());
-    assertEquals(1, result.getReply().getValue());
+    CounterTopicSubscriptionActionTestKit testKit = CounterTopicSubscriptionActionTestKit.of(CounterTopicSubscriptionAction::new);
+    // ActionResult<Empty> result = testKit.decrease(CounterTopicApi.Decreased.newBuilder()...build());
   }
+
 }
